Single / Multiplayer blackjack with basic rules
It is possible to play it as one player against the dealer or with other people against the dealer (however the UI supports 2 players). Rules were implemented according to description on [Bicyclecards - How to Play Blackjack](https://bicyclecards.com/how-to-play/blackjack/). Not all of the game options were implemented, eg there is no split, double-down or insurance and the UI lacks the logic for taking the bets in every round and it just reuses the initial bet value for each of the player